Move all files into lib directory and require them.
