+++ lib/file_modifier.rb
