+++ lib/files/.rubocop.yml
