+++ lib/files/database_cleaner.rb
