+++ lib/files/shoulda_matchers.rb
