+++ lib/gem_configurator.rb
