+++ lib/gemfile_configurator.rb
