+++ lib/initialize_git.rb
